requirements.txt
python-docx==1.1.0
pandas==2.2.2
openpyxl==3.1.2